CRANs systems sometimes behave strangely under load
So we cannot test for timings explicitly as mentioned in section 1.6 of
WRE

> It is not reasonable to test the time taken by a command: you cannot
know how fast or how heavily loaded an R platform might be. At best you
can test a ratio of times, and even that is fraught with difficulties
and not advisable: the just-in-time compiler (JIT) and the GC may
trigger at unpredictable times, following heuristics that may change
without notice.