Install self before running benchmarks
Necessary if other packages use this workflow so that the package the workflow is being run for gets installed. Those packages should Suggest bench so that it gets installed by `setup-r-dependencies`. For bench itself, this install step ensures that bench is around to run `bench::cb_*()` and the example benchmarks themselves.