Fix configuration of breaks when using recent scales package (#140)
* Fix configuration of breaks when using recent scales package.

The latest release of the scales package has added a pair of optional
arguments, `d_transform` and `d_inverse`, in the middle of the parameter
list for `trans_new` (see https://github.com/r-lib/scales/pull/341). As
a consequence, this has shifted the position of the `breaks` parameter
from 4th to 6th.

Because the `bench_time_trans` and `bench_bytes_trans` functions from
this package were passing in the breaks object positionally, this change
in the scales package means that the breaks object is now matched with
the new `d_transform` formal parameter and it ends up being ignored.

This commit changes the arguments to `trans_new` from being positional
to being named, which should be more robust to future changes to the
function.

* Tweak formatting

* NEWS bullet

---------

Co-authored-by: Davis Vaughan <davis@posit.co>